get rid of AppDatabase in favour of free functions
