--- FoodSpec/Clients/AppDatabase.swift
@@ -1,160 +1,0 @@
-import Foundation
-import GRDB
-import os
-
-/// A database of players.
-///
-/// You create an `AppDatabase` with a connection to an SQLite database
-/// (see <https://swiftpackageindex.com/groue/grdb.swift/documentation/grdb/databaseconnections>).
-///
-/// Create those connections with a configuration returned from
-/// `AppDatabase/makeConfiguration(_:)`.
-///
-/// For example:
-///
-/// ```swift
-/// // Create an in-memory AppDatabase
-/// let config = AppDatabase.makeConfiguration()
-/// let dbQueue = try DatabaseQueue(configuration: config)
-/// let appDatabase = try AppDatabase(dbQueue)
-/// ```
-struct AppDatabase {
-    /// Creates an `AppDatabase`, and makes sure the database schema
-    /// is ready.
-    ///
-    /// - important: Create the `DatabaseWriter` with a configuration
-    ///   returned by ``makeConfiguration(_:)``.
-    init(_ dbWriter: any DatabaseWriter) throws {
-        self.dbWriter = dbWriter
-        try migrator.migrate(dbWriter)
-    }
-    
-    /// Provides access to the database.
-    ///
-    /// Application can use a `DatabasePool`, while SwiftUI previews and tests
-    /// can use a fast in-memory `DatabaseQueue`.
-    ///
-    /// See <https://swiftpackageindex.com/groue/grdb.swift/documentation/grdb/databaseconnections>
-    let dbWriter: any DatabaseWriter
-}
-
-// MARK: - Database Configuration
-
-extension AppDatabase {
-    private static let sqlLogger = Logger(subsystem: Bundle.main.bundleIdentifier!, category: "GRDB")
-
-    /// Returns a database configuration suited for `PlayerRepository`.
-    ///
-    /// SQL statements are logged if the `SQL_TRACE` environment variable
-    /// is set.
-    ///
-    /// - parameter base: A base configuration.
-    public static func makeConfiguration(_ base: Configuration = Configuration()) -> Configuration {
-        var config = base
-        
-        // An opportunity to add required custom SQL functions or
-        // collations, if needed:
-        // config.prepareDatabase { db in
-        //     db.add(function: ...)
-        // }
-        
-        // Log SQL statements if the `SQL_TRACE` environment variable is set.
-        // See <https://swiftpackageindex.com/groue/grdb.swift/documentation/grdb/database/trace(options:_:)>
-        if ProcessInfo.processInfo.environment["SQL_TRACE"] != nil {
-            config.prepareDatabase { db in
-                db.trace {
-                    sqlLogger.log(level: .debug, "\($0)")
-                }
-            }
-        }
-        
-#if DEBUG
-        // Protect sensitive information by enabling verbose debugging in
-        // DEBUG builds only.
-        // See <https://swiftpackageindex.com/groue/grdb.swift/documentation/grdb/configuration/publicstatementarguments>
-        config.publicStatementArguments = true
-#endif
-        
-        return config
-    }
-}
-
-// MARK: - Database Migrations
-
-extension AppDatabase {
-    /// The DatabaseMigrator that defines the database schema.
-    ///
-    /// See <https://swiftpackageindex.com/groue/grdb.swift/documentation/grdb/migrations>
-    private var migrator: DatabaseMigrator {
-        var migrator = DatabaseMigrator()
-        
-#if DEBUG
-        // Speed up development by nuking the database when migrations change
-        // See <https://swiftpackageindex.com/groue/grdb.swift/documentation/grdb/migrations>
-        migrator.eraseDatabaseOnSchemaChange = true
-#endif
-        
-        migrator.registerMigration("createFood") { db in
-            // Create a table
-            // See <https://swiftpackageindex.com/groue/grdb.swift/documentation/grdb/databaseschema>
-            try db.create(table: "food") { t in
-                t.autoIncrementedPrimaryKey("id")
-                t.column("name", .text).notNull().unique(onConflict: .replace)
-                t.column("energy", .double)
-                t.column("fatTotal", .double)
-                t.column("fatSaturated", .double)
-                t.column("protein", .double)
-                t.column("sodium", .double)
-                t.column("potassium", .double)
-                t.column("cholesterol", .double)
-                t.column("carbohydrates", .double)
-                t.column("fiber", .double)
-                t.column("sugar", .double)
-            }
-        }
-        
-        // Migrations for future application versions will be inserted here:
-        // migrator.registerMigration(...) { db in
-        //     ...
-        // }
-        
-        return migrator
-    }
-}
-
-// MARK: - Database Access: Writes
-// The write methods execute invariant-preserving database transactions.
-
-extension AppDatabase {
-    /// Saves (inserts or updates) a player. When the method returns, the
-    /// player is present in the database, and its id is not nil.
-    func saveFood(_ food: inout Food) async throws {
-        food = try await dbWriter.write { [food] db in
-            try food.saved(db)
-        }
-    }
-    
-    /// Delete the specified players
-    func deleteFoods(ids: [Int64]) async throws {
-        try await dbWriter.write { db in
-            _ = try Food.deleteAll(db, keys: ids)
-        }
-    }
-    
-    /// Delete all players
-    func deleteAllFoods() async throws {
-        try await dbWriter.write { db in
-            _ = try Food.deleteAll(db)
-        }
-    }
-}
-
-extension Column: Hashable {
-    public static func == (lhs: Column, rhs: Column) -> Bool {
-        lhs.name == rhs.name
-    }
-
-    public func hash(into hasher: inout Hasher) {
-        hasher.combine(name)
-    }
-}
