+++ FoodSpec/Clients/Database/DatabaseClient.swift
@@ -1,5 +1,5 @@
 //
-//  GRDBDatabaseClient.swift
+//  DatabaseClient.swift
 //  FoodSpec
 //
 //  Created by Victor Socaciu on 05/12/2023.
@@ -10,15 +10,15 @@ import GRDB
 import ComposableArchitecture
 
 @DependencyClient
-struct GRDBDatabaseClient {
+struct DatabaseClient {
     var getRecentFoods: (_ sortedBy: Food.SortingStrategy, _ order: SortOrder) async throws -> [Food]
     var insert: (_ food: Food) async throws -> Food
     var delete: (_ food: Food) async throws -> Void
 }
 
-extension GRDBDatabaseClient: DependencyKey {
-    static var liveValue: GRDBDatabaseClient = {
-        let db = AppDatabase.shared.dbWriter
+extension DatabaseClient: DependencyKey {
+    static var liveValue: DatabaseClient = {
+        let db = createAppDatabase()
         return .init(
             getRecentFoods: { sortedBy, order in
                 let column = sortedBy.column
@@ -41,12 +41,12 @@ extension GRDBDatabaseClient: DependencyKey {
         )
     }()
 
-    static let testValue: GRDBDatabaseClient = .init()
+    static let testValue: DatabaseClient = .init()
 }
 
 extension DependencyValues {
-    var grdbDatabaseClient: GRDBDatabaseClient {
-        get { self[GRDBDatabaseClient.self] }
-        set { self[GRDBDatabaseClient.self] = newValue }
+    var databaseClient: DatabaseClient {
+        get { self[DatabaseClient.self] }
+        set { self[DatabaseClient.self] = newValue }
     }
 }
